Rewritten introduction, renamed ebook, and removed data versioning chapter and a few scoping figures.
The introduction is now in the scientific, implicit format, and incorporates the insights stage.
The ebook was renamed to "Navigating Machine Learning Projects" which is shorter than "Decision Making in Machine Learning Projects" and sounds better. Additionally, it matches nicely with the ML compass as the title image.
The data versioning chapter is my personal strategy, but probably not ideal for real word business problems, that rather rely on more professional dataset version control. Additionally, it is not really a decision, but rather a method of working in an organized way. I could also make a chapter about tidying up the desk in the office, which probably also is a success factor in ML projects, but not the clear focus of this book.
Some figures of the scoping stage where just the informal and hand drawn screenshots from Andrew Ngs lectures. I removed them, because the info anyways is in the text.