Reviewed scoping process chapter and added table to Data Centric Modeling chapter
Added info on when to use ML and when not to use it.
Added a table with the info from Andrew Ng on three tasks where improving the data yielded a considerably higher performance gain than tuning the hyperparameters/working on improving the model.

Overall added smaller refinements to the document, such as spacing before subsections, sections, etc.